refactor(zkvm): move e2e tests out of prover
